pyextract: add get stack information from logging message
stack_info, which defaults to False. If true, stack information is added to the logging
message, including the actual logging call.

Note that this is not the same stack information as that displayed through specifying
exc_info: The former is stack frames from the bottom of the stack up to the logging call
in the current thread, whereas the latter is information about stack frames which have
been unwound, following an exception, while searching for exception handlers.

reference:
- https://docs.python.org/3/library/logging.html#logging.Logger.debug

Signed-off-by: Junbo Zheng <zhengjunbo1@xiaomi.com>